requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-telegram-bot==21.0
google-generativeai==0.8.3
httpx==0.27.0
google-cloud-firestore==2.16.0
pydantic-settings==2.4.0
google-cloud-logging==3.10.0
google-auth==2.29.0
